backend/requirements.txt
# ─── Web framework ────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# ─── Config & validation ──────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ─── LLM — Claude ─────────────────────────────────────────────
anthropic==0.40.0

# ─── Embeddings — local (no API key required) ────────────────
sentence-transformers

# ─── Vector database ──────────────────────────────────────────
chromadb==0.5.23

# ─── Document parsing ─────────────────────────────────────────
pypdf==5.1.0
python-docx==1.1.2
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# ─── Text processing ──────────────────────────────────────────
tiktoken==0.8.0
nltk==3.9.1

# ─── HTTP client ──────────────────────────────────────────────
httpx==0.28.1
